+++ cmd/local/conf/conf.go
@@ -358,7 +358,7 @@ func ReadConf(overrides map[string]string, ddcYamlLoc string) (*CollectConf, err
 				var err error
 				detectedConfig, err = GetConfiguredDremioValuesFromPID(c.dremioPID)
 				if err != nil {
-					msg := fmt.Sprintf("unable to retrieve configuration from pid %v: %v", c.dremioPID, err)
+					msg := fmt.Sprintf("AUTODETECTION DISABLED: will rely on ddc.yaml configuration as ddc is unable to retrieve configuration from pid %v: %v", c.dremioPID, err)
 					fmt.Println(msg)
 					simplelog.Errorf(msg)
 				} else {
@@ -366,7 +366,7 @@ func ReadConf(overrides map[string]string, ddcYamlLoc string) (*CollectConf, err
 					c.dremioConfDir = detectedConfig.ConfDir
 				}
 			} else {
-				fmt.Println("no valid pid found therefore the log and configuration autodetection will not function")
+				fmt.Println("AUTODETECTION DISABLED: will rely on ddc.yaml configuration as the ddc user does not have permissions to the dremio process consider using --sudo-user to resovle this")
 				simplelog.Warning("no valid pid found therefor the log and configuration autodetection will not function")
 			}